+++ mypy/messages.py
@@ -459,7 +459,7 @@ class MessageBuilder:
         elif member == "__call__":
             if isinstance(original_type, Instance) and (
                 original_type.type.fullname
-                in {"typing._Callable", "typing._NamedCallable", "builtins.function"}
+                in {"typing.Callable", "typing._NamedCallable", "builtins.function"}
             ):
                 # "'function' not callable" is a confusing error message.
                 # Explain that the problem is that the type of the function is not known.