+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - check type args on type aliases (#817)
 - cache modules that only have baseline errors
 - fix dmypy in vscode
+- fix `typing.Callable`
 ### Docs
 - correct error code documentation regarding defaults